Refactor species exports to direct naming and enhance diagnostics
- Removed 'total_' prefix and '_emissions' suffix from ESMF exports.
- Species defined in the 'species' block are now exported directly (e.g., 'nox', 'co').
- Physics schemes (Lightning, SoilNox, Dust, Megan, DMS, Volcano) now export their specific components separately as independent ESMF fields (e.g., 'lightning_nox', 'soil_nox').
- Enhanced AcesDiagnosticManager to support CF-compliant metadata (units, long_name).
- Improved WriteDiagnostics to safely handle custom diagnostics using a data-copying template mechanism.
- Fixed codebase-wide linting violations: added explicit braces to all control blocks and resolved narrowing conversion warnings with static_cast.
- Refined CMake configuration for robust Fortran module handling in CI environments.

+++ src/physics/aces_dms_fortran_bridge.cpp
@@ -41,9 +41,9 @@ void DMSFortranScheme::Run(AcesImportState& import_state, AcesExportState& expor
     dv_seaconc.sync<Kokkos::HostSpace>();
     dv_dms_emis.sync<Kokkos::HostSpace>();
 
-    int nx = dv_dms_emis.extent(0);
-    int ny = dv_dms_emis.extent(1);
-    int nz = dv_dms_emis.extent(2);
+    int nx = static_cast<int>(dv_dms_emis.extent(0));
+    int ny = static_cast<int>(dv_dms_emis.extent(1));
+    int nz = static_cast<int>(dv_dms_emis.extent(2));
 
     run_dms_fortran(dv_u10.view_host().data(), dv_tskin.view_host().data(),
                     dv_seaconc.view_host().data(), dv_dms_emis.view_host().data(), nx, ny, nz);
